software/Web Server/EB Application/requirements.txt
Flask==0.9
xively-python==0.1.0-rc0
numpy==1.8.1
scipy==0.10.1